feat(formatting): disable default tsserver formatting in favor of efm
Disable tsserver formatting since it conflicts with preferred formatting
by prettier based on eslint rules. Currently this isn't based on the
presence of a config so even projects that don't contain prettier will
be formatted. Automatic formatting can be temporarily disabled for now
when not desired. Use synchronous formatting on save to be sure updates
made before a given file is saved.